move torch_dtype_from_precision to llm.utils
